Create README.md
Conjugator
This is a personal project that has like objective to help to learn the verbs conjugations of a determined language. In this case is done to learn french verbs. It is programmed in Python and Flask and also there is a Java version to practice and to improve the coding in these programming languages. This is the fisrt version of this project, in the future it will added new features.